documentación y revisión del ejemplo para validar la suma
