+++ nbproject/private/private.xml
@@ -2,9 +2,6 @@
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
     <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/G:/Mi%20unidad/Logica%20y%20representacion%20III/desarrollo/Polinomios/src/polinomios/listas/simple/PolinomioListaSimple.java</file>
-            <file>file:/G:/Mi%20unidad/Logica%20y%20representacion%20III/desarrollo/Polinomios/src/polinomios/listas/simple/EjemploLSL.java</file>
-        </group>
+        <group/>
     </open-files>
 </project-private>
