+++ src/polinomios/listas/simple/PolinomioListaSimple.java
@@ -22,10 +22,10 @@
  */
 package polinomios.listas.simple;
 
-
 import polinomios.util.Termino;
 
 /**
+ * Clase que representa un polinomio en lista ligada simple con nodo cabeza
  *
  * @author alejandroescobar
  */
@@ -33,11 +33,19 @@ public class PolinomioListaSimple {
 
     private final Nodo cabeza;
 
+    /**
+     * Constructor para un polinomio sin terminos
+     */
     PolinomioListaSimple() {
         cabeza = new Nodo(null);
         cabeza.setLiga(null);
     }
 
+    /**
+     * Obtener el nodo cabeza de la lista
+     *
+     * @return
+     */
     public Nodo getCabeza() {
         return cabeza;
     }
@@ -56,17 +64,27 @@ public class PolinomioListaSimple {
     public int getGrado() {
         Nodo primero = cabeza.getLiga();
         return primero.getTermino().getE();
+        // ToDo - Validación pendiente por realizar.
     }
 
-
+    /**
+     * Metodo para sumar dos polinomios, suma el polinomio a con el polinomio b
+     * que se pasa como parametros, no se modifica el polinomio a.
+     *
+     * @param pB polinomio a sumar
+     * @return polinomio con el resultado de la suma.
+     */
     public PolinomioListaSimple sumar(PolinomioListaSimple pB) {
         PolinomioListaSimple pC = new PolinomioListaSimple();
 
+        // Variables para los recorridos en polinomios a,b y c
         Nodo prA = this.getCabeza().getLiga();
         Nodo prB = this.getCabeza().getLiga();
         Nodo uC = pC.getCabeza();
 
+        // Recorrido de los terminos en alguna de las dos listas
         while (!finRecorrido(prA) && !finRecorrido(prB)) {
+            // Variables de exponente y coeficiente de terminos
             int eA = prA.getTermino().getE();
             double cA = prA.getTermino().getC();
             int eB = prB.getTermino().getE();
@@ -75,58 +93,64 @@ public class PolinomioListaSimple {
             if (eA > eB) { //Traslado datos del Termino de A.
                 Termino t = new Termino(eA, cA);
                 Nodo n = new Nodo(t);
-                prA = prA.getLiga();
                 uC.setLiga(n);
                 uC = n;
-            } else if (eA == eB) {
-                double pcC = cA + cB;
-                if (pcC != 0) {
-                    Termino t = new Termino(eA, pcC);
+                prA = prA.getLiga(); // recorro la lista polinomio de a
+            } else if (eA == eB) { // Opero terminos de a y b
+                double posibleCoeficienteC = cA + cB;
+                // Si el coeficiente resultado es 0 no se lleva a polinomio c.
+                if (posibleCoeficienteC != 0) {
+                    Termino t = new Termino(eA, posibleCoeficienteC);
                     Nodo n = new Nodo(t);
                     uC.setLiga(n);
                     uC = n;
                 }
+                // recorro la lista de polinomio a y b
                 prA = prA.getLiga();
                 prB = prB.getLiga();
             } else {
                 Termino t = new Termino(eB, cB);
                 Nodo n = new Nodo(t);
-
-                prB = prB.getLiga();
                 uC.setLiga(n);
                 uC = n;
+                prB = prB.getLiga(); // recorro la lista polinomio de b
             }
-
         }
 
+        // Si la lista de terminos del polinomio b finalizo primero
         while (!finRecorrido(prA)) {
+            // Variables para exponente y coefiente de lista polinomio a
             int eA = prA.getTermino().getE();
             double cA = prA.getTermino().getC();
+            // Creo el nuevo termino, nuevo nodo y ligo en c
             Termino t = new Termino(eA, cA);
             Nodo n = new Nodo(t);
             uC.setLiga(n);
             uC = n;
+            // me sigo moviendo en lista a
             prA = prA.getLiga();
         }
 
+        // Si la lista de terminos del polinomio a finalizo primero
         while (!finRecorrido(prB)) {
+            // Variables para exponente y coefiente de lista polinomio a
             int eB = prB.getTermino().getE();
             double cB = prB.getTermino().getC();
+            // Creo el nuevo termino, nuevo nodo y ligo en c
             Termino t = new Termino(eB, cB);
             Nodo n = new Nodo(t);
             uC.setLiga(n);
             uC = n;
+            // me sigo moviendo en lista b
             prB = prB.getLiga();
         }
 
         return pC;
     }
 
-
     public int getCoeficiente(int e) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
     }
-
 
     public int getDiferentesCero() {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
@@ -135,7 +159,6 @@ public class PolinomioListaSimple {
     private boolean finRecorrido(Nodo p) {
         return p == null;
     }
-
 
     public void ingresar(double coeficiente, int exponente) {
         throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
